refactor: move database health logic from controller to service
